Bug fixed : THRD_localShift is a double instead of integer otherwise, problem with the selection of the reference at Antarctica
git-svn-id: http://svn-ae/ae_doas/qdoas/trunk/Src@1759 0551a244-8104-0410-bdc1-badfa85f4f5d

+++ engine/winthrd.h
@@ -9,7 +9,7 @@ extern unsigned int THRD_id;                   // thread identification number
 extern int       THRD_levelMax;                // level of thread
 extern int       THRD_lastEvent;               // last event
 extern unsigned long THRD_delay;               // wait for next event
-extern int       THRD_localShift;
+extern double    THRD_localShift;
 extern int       THRD_correction;
 extern int       THRD_browseType;
 extern int       THRD_treeCallFlag;
